Add in fleecing fixture to cfme.fixtures.appliance
